Add files via upload
The code here only works on a certain array, which is [6,5,2,4,7]. The code created is not yet to be robust, but I am thinking of a way to enhance the code. If anyone knows how to improve the code; please feel free to create a pull request. In advance, i will give an instruction to the objective of the code. 

How do you find the largest and smallest number in an unsorted integer array? --> [6,5,2,4,7]: Organize the array, the code is based on this question. 

How I answered the question --> 1). I created a function to arrange the numbers in the array
                                                       2). I created another function to find the smallest and the largest number in the array. 
                                                       3). Called the two functions